Move factory methods to IOUringHandler
Motivation:

In all our other transport implementations we have the factory methods in the IoHandler implementation itself. Let's be consistent

Modifications:

Move methods from IOUring to IOUringHandler.

Result:

More consistent API
